Added a secondary round of writing to the audit log. If in the rather daft situation (and CLUK have reported it!) whereby they have an audited column which relies on a calculation which uses the audit functionality (field changed between two dates, field last changed date functions) then  it was not being populated. Long story short, this is a chicken and egg situation, the calculations must fire first so that we can then populate the audit log, but the calculations could be wrong if they are reliant on the audit log being needed. So we refire calcs after an audit where there's calcs that rely on the audit. And now (and this is the changed bit) we also look at these calcs and place something in the audit for that.
Yes - I know you could start setting up even more complex levels, or start hacking round with inserts and deletes. This won't fix that - all that's going to be resolved is the issue reported.

The real fix (nexus onwards) would be to say that you can't audit calculations - why should it be necessary - you should only want to audit data that a user has changed rather than anything that is system generated. 


Note: Curious fun with removing the _ appending from some of the code since moving to VS2015. This should have no bearing, but we possibly need to keep and eye on it if we're compiling with an older version.