Added an option to define a webform screen as requiring authentication before loading (this is just the definition saving - actual implementation is a different changeset)
The option for authorisation should only appear on web form properties and no other controls. 

Because we need an extra property on the webform I've had to recimpile that whole activeX - I looked at merging into the project a la stored data elements, but I hit all sorts of squiffinesses because of click events and the validation storing the objects in arrays, so I left well alone. So mkake sure you get that whole project and recompile on your PC before running this - otherwise it WILL crash.