Swap the ASRSysColumns, ASRSysViews, ASRSysTables, ASRSysWorkflows from views back to tables. Details of why are in the associated task - but briefly its because DAO is rubbish and does not map to views very well.
So in fixing I have come to the conclusion that the system manager save is a bit of a mess. Sometimes we insert directly into the tables by executing strings (tables & views), sometimes we scoll through properties and update the recordset object (columns and workflows). Where we assign a new Id is also a bit wibbly-wobbly.

While I'm ranting VB6 doesn't like guids, so I've had to pull in an API to do this. 

I've removed removed the hack that ignored the columns not on a base table (locked, guid). The last updated column has been removed because it was showing erroneous results as are constantly regenerating tables and columns even if there are no changes.

Reflected these changes in the database project by deleting the triggers from these views, removing the view themselves, renaming the tables, and finally adding locked and guid to the tables.