Ensure that the no template is uploaded is recorded in the event log.
You'll not there's a bit of duplication in the error reporting bit with one part looking for a error message in a string and another in an array. Ideally this would be one type of collection, but the rest of the merge engine is a bit confused and I'm abit scared to sort that out right now.