This is the converted match report engine from vb6 Data Manager (with some modifications). I'm checking in separently to other work required to actually call it (give me half hour), but this will need reviewing in conjunction with other the other talent report tasks.
I've gone through the conversion process in a similar mannner to how DataTransferRun was converted.
1) Auto convert code using Visual Studio 2008 project upgrade tool.
2) Copy relevant files into VS2015 project.
3) Swap out ADO for ADO.NET
4) Resize arrays
5) Diagnostic and fix errors
6) Add SecondTable code to the expression builder (when the expression builder was originally ported into server dll match reports had not been developed - so I've had to add in these changes)
7) Extend the DisplayColumn as these were originally properties from the activeX grid

I have used the word "hack" in a few places as there are some globals required. When the whole Talent Management engine is running there will be a refractoring stage to tidy these up - but please feel free to comment on them, I'm sure there's more...