Added some of the trimmings to call the recently migrated match report engine. Defsel for match reports has been included for testing the engine purposes however it is not envisionaged that this menu item will survive subsequent builds once the talent management definition screens have been added and some data populated.
These reports are available via the accordian and should also appear in the find window and record edit screens (if on the appropraite table)

Notes:
Ensure that you have run ALL the latest update scripts.
The icon for calendar reports has been used for the moment - once people can decide what actual icon to use this will be updated.
The first time a talen/match report is run the talent barcharty does not seem to render quite right (again this may be replaced with another method of display, so not urgent to panic about this)