Standardised the login failed error message to make it oWasp compliant (if not terribly useful for the user - but then that would also be useful for any potential hacker - damn those pesy hackers!)
Now there is one slight quirk with this in that if you create a new user and set it to force change password on first login and also that security group has no module access then we have to go through the password malarkey before we can test if they have that module access becuase we can't establish a db connection until the password has been changed. So that will go through ok and allow them to change it, but then when actually go through the proper login it will return a login failed message. All quite valid, but might be a bit confusing for the user.