The talent report is now themerollered. Changed the control from the sparkle to a stylesheet. Fun and games ensued here trying to resize columns as the only way you'll know the width of a grid is to set is columns sizes, and you can't work out columns sizes as a percentage until you've worked out the size of the grid. Add this to the fact that it doesn't always snap to the grid and that if you put autosize on then you can't adjust teh column width. Really we should be calculating the size using tyhe colModel, but this would involve changing the generic jqGridHelper and that's just too scary for me.
Themeroller bits
Used the caption background for the actual bar, minimum and prefered backgrounds use state and the vertical minimum and prefered use error state.

Printing is an issue with this as the way we print is lopping of the last columns (this is a problems for custom reports too). This will have to be addressed in a separate changeset.