Reverted the change in the assembly to the previous one (was in the midddle of other fixes which should never have gone in - will break some sql versions). Then rolled the int check changes for the sql version in number into actualuserdetails
The udfASRSQLVersion is regenerated in the system manager (I suspect this is only to handle existing upgrade paths where we have a chicken and egg situation).

Please note running the web update script as it stands at the moment causes an error with the viewid on the orgcharts. This is nothing to do with this fix and needs remedying seperately (and should never have been checked in in that state!)