Error dropping control in workflow designer for multi-line fields.
Sorry, but vb6 and visual studio for some reason occasionally decide to change case of variables. tabPages -> tabPages. This is not of my doing. The only code change in here to be aware of is on about line 5787

pctlControl.Width = TextWidth(String(500, "W")) + (2 * XFrame).

Basically the 8000 is too wide for the new default font (Segoe UI)