User Story : 23729 - Unique code changes. We are now using the SQL sequence object to implement the unique code functionality. Basically this function is a automatically updating numeric code. Note: this is not just a simple implementation of this sequence, we have to wrap this code around some of our own special code because the OpenHR triggers and calculations embed duplicate references to this calc which would cause the sequence to jump intermittently.
Basic process is as follows (same basic logic as before this rework - just altered to use sequence object)
A record is updated or inserted.
The unique code bucket is flushed. (spstat_flushuniquecode)
All calls to the uniquecode function (udfstat_getuniquecode) loop through previously allocated numbers for that transaction - if not allocated gave root value.
The unique code bucket is flushed again. This will generate any sequences if necessary.

Things to note.
Sequences have to be created at runtime instead of in the system manager as the root code value could in fact be a column of data and different for every record (see user story for examples)
The update script will take the existing tbsys_uniquecode table and generate sequences for them - meaning a seamless upgrade.
Getting the server assembly debug running via Visual Studio is a pain, if you value your day, just compile as if for a typical build. (i.e. publish to a server of you choice and copy the generated assembly code into the assembly sql script - then shift login to your db)
The .NET framework referenced has had to be upgraded to 4.61 get Visual Studio to compile against SQL2012 (a pre-requisit for the SEQUENCE functionality)
The added script.PreDeployment is only used if you are attempting to run the assembly in debug mode (good luck!)