Bug 25065 - The sort order was not getting evaluated on reports. In a previous version of SQL Microsoft chnaged the way the insert into clause worked so that an ORDER BY would not get evaluated (makes sense if you think about it as why should you care about the physical order of records - you only care when you come to read them, ditto the same for handling clustered indexes, that would mess things up anyway -  I digress).
Anyhow, we got around this by putting a TOP 1 trillion into the clause which seemed to trick sql into actually doing the ordering. But as we now have discovered this seems to stop working when the dataset gets a bit larger (either through a large amount of records, or a lot of columns, or a combo of both). On a simple standard system I replicated this by reporting from absence as a base table, having one column (start date) and hoofing in about 46 thousand records. At this level is when it started to go funny, adding some columns brought that threshold down a bit. Curiously its doesn't have to actually return that amount of rows, its just that its somewhere in the first INSERT INTO query has to evalaute a large dataset.

So (if you're still reading) the solution is to move where the ORDER BY clause is evaluated from the inserting into the temporary table and to the location where the data is read back and processed by the client.

While only reported in custom reports in the DM this would also affect the web and also exporting in the DM (same code fix - different files).

Things to note: Ordering works slightly different when you have child tables in play, it fakes a different order and uses this rather than the mstrSQLOrderBy. This would have previously worked as far as I can tell and, so I haven't changed this extra logic, it will have been affected by the change in the basic code.  It does get complicated in multi-child tables with the breaks and the suppress repeated options! 

Also to note that this code is shared by the bradford factor, which also wasn't affected because it handled its own orders, but the logic has had to be altered to cater for the new basic logic of the custom report engine.

The original bug was reported on a sig database "[S] Training Delegate List" and was logged in a user with access to the base tables. Once again the code thats built up can differ depending on the security of the user - someone without access to the base table will have many more joins. I mention this, not because I've changed any of the logic, but because that could have affected the size of the original dataset and may have made the original problem manifest itself differently - I've sanity checked this, but if you've got time on your hands... :-)

Almost done... Sig reported that this has happened since an upgrade. Mmmmm, this code has been like this for many versions, so the "upgrade" is either a change of sql version, a possible sql hotfix that Microsoft occasionally deploy (although I have no evidence of this) or possibly a change to the report thats added some extra columns thats meant the dataset being processed has been largified.

Right, I'm going to copy these comments into the associated work item. Thanks for listening - have a nice day.